misc: move folder structure to fix npm script
