+++ package.json
@@ -7,7 +7,7 @@
     "chain": "hardhat node",
     "fork": "hardhat node --fork https://mainnet.infura.io/v3/460f40a260564ac4a4f4b3fffb032dad",
     "test:fork": "FORK=true hardhat test --network hardhat ./test/mainnet-fork/*.ts",
-    "test": "FORK=false hardhat test --network hardhat ./test/unit-tests/*.ts",
+    "test": "FORK=false hardhat test --network hardhat ./test/unit-tests/**/*.ts",
     "coverage": "npx hardhat coverage --testfiles 'test/unit-tests/**/*.ts'",
     "compile": "hardhat compile",
     "lint:contract": "npx hardhat format",
