+++ test/unit-tests/action-examples/SellAction.ts
@@ -1,10 +1,10 @@
 import { ethers, waffle } from "hardhat";
-import { BigNumber } from "ethers";
+import { BigNumber, utils } from "ethers";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { expect } from "chai";
-import { getOrder } from "../utils/orders";
+import { getOrder } from "../../utils/orders";
 import {
-  LongOToken,
+  ShortOToken,
   MockERC20,
   MockWhitelist,
   MockSwap,
@@ -13,8 +13,9 @@ import {
   MockOToken,
   MockOpynOracle,
   MockEasyAuction,
-} from "../../typechain";
+} from "../../../typechain";
 import * as fs from "fs";
+import { parseUnits } from "@ethersproject/units";
 
 const mnemonic = fs.existsSync(".secret")
   ? fs.readFileSync(".secret").toString().trim()
@@ -26,14 +27,14 @@ enum ActionState {
   Activated,
 }
 
-describe("LongAction: Buying Puts", function () {
+describe("ShortAction", function () {
   const provider = waffle.provider;
 
   const counterpartyWallet = ethers.Wallet.fromMnemonic(mnemonic, "m/44'/60'/0'/0/30");
 
-  let action: LongOToken;
+  let action: ShortOToken;
   // asset used by this action: in this case, weth
-  let weth: MockERC20;
+  let token: MockERC20;
   //
   let usdc: MockERC20;
 
@@ -54,10 +55,9 @@ describe("LongAction: Buying Puts", function () {
   let otoken1: MockOToken;
   let otoken2: MockOToken;
 
-  const wethPrice = 2500 * 1e8;
-  const otokenBadStrikePrice = 2400 * 1e8; // strike too high
-  const otoken1StrikePrice = 2000 * 1e8;
-  const otoken2StrikePrice = 2100 * 1e8;
+  const otokenBadStrikePrice = 10 * 1e8;
+  const otoken1StrikePrice = 4000 * 1e8; // 4000
+  const otoken2StrikePrice = 5000 * 1e8; // 5000
 
   let otoken1Expiry = BigNumber.from(0);
   let otoken2Expiry = BigNumber.from(0);
@@ -84,8 +84,8 @@ describe("LongAction: Buying Puts", function () {
 
   this.beforeAll("Deploy Mock contracts", async () => {
     const ERC20 = await ethers.getContractFactory("MockERC20");
-    weth = (await ERC20.deploy()) as MockERC20;
-    await weth.init("WETH", "WETH", 18);
+    token = (await ERC20.deploy()) as MockERC20;
+    await token.init("WETH", "WETH", 18);
 
     usdc = (await ERC20.deploy()) as MockERC20;
     await usdc.init("USDC", "USDC", 6);
@@ -116,20 +116,24 @@ describe("LongAction: Buying Puts", function () {
 
   describe("deployment test", () => {
     it("deploy", async () => {
-      const LongActionContract = await ethers.getContractFactory("LongOToken");
-      action = (await LongActionContract.deploy(
+      const ShortActionContract = await ethers.getContractFactory("ShortOToken");
+      action = (await ShortActionContract.deploy(
         vault.address,
-        usdc.address,
+        token.address,
         swap.address,
         auction.address,
-        controller.address
-      )) as LongOToken;
+        controller.address,
+        0 // type 0 vault
+      )) as ShortOToken;
 
       expect((await action.owner()) == owner.address).to.be.true;
 
-      expect((await action.asset()) === usdc.address).to.be.true;
+      expect((await action.asset()) === token.address).to.be.true;
 
-      expect((await usdc.allowance(action.address, vault.address)).eq(ethers.constants.MaxUint256)).to.be.true;
+      expect(await controller.vaultOpened()).to.be.true;
+
+      expect((await token.allowance(action.address, pool.address)).eq(ethers.constants.MaxUint256)).to.be.true;
+      expect((await token.allowance(action.address, vault.address)).eq(ethers.constants.MaxUint256)).to.be.true;
 
       // init state should be idle
       expect((await action.state()) === ActionState.Idle).to.be.true;
@@ -137,67 +141,83 @@ describe("LongAction: Buying Puts", function () {
       // whitelist is set
       expect((await action.opynWhitelist()) === whitelist.address).to.be.true;
     });
+    it("should deploy with type 1 vault", async () => {
+      const ShortActionContract = await ethers.getContractFactory("ShortOToken");
+      await ShortActionContract.deploy(
+        vault.address,
+        token.address,
+        swap.address,
+        ethers.constants.AddressZero,
+        controller.address,
+        1 // type 0 vault
+      );
+      expect((await action.owner()) == owner.address).to.be.true;
+      expect((await action.asset()) === token.address).to.be.true;
+      expect(await controller.vaultOpened()).to.be.true;
+    });
   });
 
-  const totalDepositInAction = 100000 * 1e6;
+  const totalDepositInAction = utils.parseUnits("100");
 
   describe("idle phase", () => {
-    before("Mint some usdc to action", async () => {
-      // mint 100000 usdc
-      await usdc.mint(action.address, totalDepositInAction);
+    before("Mint some eth to action", async () => {
+      // mint 100 weth
+      await token.mint(action.address, totalDepositInAction);
     });
     before("Deploy mock otokens", async () => {
       const MockOToken = await ethers.getContractFactory("MockOToken");
       otoken1 = (await MockOToken.deploy()) as MockOToken;
-      await otoken1.init("oWETHUSDC-P", "oWETHUSDC-P", 18);
+      await otoken1.init("oWETHUSDC", "oWETHUSDC", 18);
       await otoken1.initMockOTokenDetail(
-        weth.address,
+        token.address,
         usdc.address,
-        usdc.address,
+        token.address,
         otoken1StrikePrice,
         otoken1Expiry,
-        true
+        false
       );
 
       otoken2 = (await MockOToken.deploy()) as MockOToken;
-      await otoken2.init("oWETHUSDC-P", "oWETHUSDC-P", 18);
+      await otoken2.init("oWETHUSDC", "oWETHUSDC", 18);
       await otoken2.initMockOTokenDetail(
-        weth.address,
+        token.address,
         usdc.address,
-        usdc.address,
+        token.address,
         otoken2StrikePrice,
         otoken2Expiry,
-        true
+        false
       );
 
       otokenBad = (await MockOToken.deploy()) as MockOToken;
       await otokenBad.init("oWETHUSDC", "oWETHUSDC", 18);
       await otokenBad.initMockOTokenDetail(
-        weth.address,
+        token.address,
         usdc.address,
-        usdc.address,
+        token.address,
         otokenBadStrikePrice,
         otoken2Expiry,
-        true
+        false
       );
 
-      await oracle.setAssetPrice(weth.address, wethPrice); // 1000 USD
+      await oracle.setAssetPrice(token.address, 10000000000);
     });
-    it("should revert if calling trade in idle phase", async () => {
-      const premium = 5000 * 1e6;
-      const buyAmount = 20 * 1e8;
-
+    it("should revert if calling mint + sell in idle phase", async () => {
+      const collateral = utils.parseUnits("10");
+      const amountOTokenToMint = 10 * 1e8;
+      const premium = parseUnits("1");
       const order = await getOrder(
         action.address,
-        usdc.address,
-        premium,
-        counterpartyWallet.address,
         otoken1.address,
-        buyAmount,
+        amountOTokenToMint,
+        counterpartyWallet.address,
+        token.address,
+        premium.toString(),
         swap.address,
         counterpartyWallet.privateKey
       );
-      await expect(action.connect(owner).tradeAirswapOTC(order)).to.be.revertedWith("!Activated");
+      await expect(
+        action.connect(owner).mintAndTradeAirSwapOTC(collateral, amountOTokenToMint, order)
+      ).to.be.revertedWith("!Activated");
     });
     it("should not be able to token with invalid strike price", async () => {
       await expect(action.connect(owner).commitOToken(otokenBad.address)).to.be.revertedWith("Bad Strike Price");
@@ -213,6 +233,7 @@ describe("LongAction: Buying Puts", function () {
   });
 
   describe("activating the action", () => {
+    const mintOTokenAmount = 10 * 1e8;
     before("increase blocktime to get it over with minimal commit period", async () => {
       const minPeriod = await action.MIN_COMMIT_PERIOD();
       await provider.send("evm_increaseTime", [minPeriod.toNumber()]); // increase time
@@ -229,107 +250,158 @@ describe("LongAction: Buying Puts", function () {
     it("should get currentValue as total amount in gamma as ", async () => {
       expect((await action.currentValue()).eq(totalDepositInAction)).to.be.true;
     });
-    describe("long with AirSwap", async () => {
-      before("mint some option ", async () => {
-        await otoken1.connect(owner).mint(counterpartyWallet.address, 20 * 1e8);
-      });
-      it("should be able to buy in this phase", async () => {
-        const premium = 5000 * 1e6;
-        const buyAmount = 20 * 1e8;
+    describe("short with AirSwap", async () => {
+      it("should not be able to mint and sell if less than min premium", async () => {
+        const collateralAmount = utils.parseUnits("10");
+        const sellAmount = 10 * 1e8;
+        const premium = utils.parseUnits("0");
         const order = await getOrder(
           action.address,
-          usdc.address,
-          premium,
-          counterpartyWallet.address,
           otoken1.address,
-          buyAmount,
+          sellAmount,
+          counterpartyWallet.address,
+          token.address,
+          premium.toString(),
           swap.address,
           counterpartyWallet.privateKey
         );
-
-        const usdcBalanceBefore = await usdc.balanceOf(action.address);
-        await action.connect(owner).tradeAirswapOTC(order);
-        const usdcBalanceAfter = await usdc.balanceOf(action.address);
-        expect(usdcBalanceBefore.sub(usdcBalanceAfter).eq(premium)).to.be.true;
+        await expect(
+          action.connect(owner).mintAndTradeAirSwapOTC(collateralAmount, mintOTokenAmount, order)
+        ).revertedWith("Need minimum option premium");
+      });
+      it("should be able to mint and sell in this phase", async () => {
+        const collateralAmount = utils.parseUnits("10");
+        const otokenBalanceBefore = await otoken1.balanceOf(action.address);
+        const sellAmount = 10 * 1e8;
+        const premium = utils.parseUnits("1");
+        const order = await getOrder(
+          action.address,
+          otoken1.address,
+          sellAmount,
+          counterpartyWallet.address,
+          token.address,
+          premium.toString(),
+          swap.address,
+          counterpartyWallet.privateKey
+        );
+        await action.connect(owner).mintAndTradeAirSwapOTC(collateralAmount, mintOTokenAmount, order);
+        const otokenBalanceAfter = await otoken1.balanceOf(action.address);
+        expect(otokenBalanceAfter.sub(otokenBalanceBefore).eq("0")).to.be.true;
       });
       it("should revert when trying to fill wrong order", async () => {
-        const premium = 5000 * 1e6;
-        const buyAmount = 20 * 1e8;
+        const collateralAmount = utils.parseUnits("10");
         const badOrder1 = await getOrder(
           action.address,
-          weth.address, // this is wong
-          premium,
+          ethers.constants.AddressZero,
+          mintOTokenAmount,
           counterpartyWallet.address,
-          otoken1.address,
-          buyAmount,
+          token.address,
+          "1",
           swap.address,
           counterpartyWallet.privateKey
         );
-        await expect(action.connect(owner).tradeAirswapOTC(badOrder1)).to.be.revertedWith("Can only pay with asset");
+        await expect(
+          action.connect(owner).mintAndTradeAirSwapOTC(collateralAmount, mintOTokenAmount, badOrder1)
+        ).to.be.revertedWith("Can only sell otoken");
 
         const badOrder2 = await getOrder(
           action.address,
-          usdc.address,
-          premium,
+          otoken1.address,
+          mintOTokenAmount,
           counterpartyWallet.address,
-          otoken2.address, // this is wrong
-          buyAmount,
+          ethers.constants.AddressZero,
+          "1",
           swap.address,
           counterpartyWallet.privateKey
         );
-        await expect(action.connect(owner).tradeAirswapOTC(badOrder2)).to.be.revertedWith("Can only buy otoken");
+        await expect(
+          action.connect(owner).mintAndTradeAirSwapOTC(collateralAmount, mintOTokenAmount, badOrder2)
+        ).to.be.revertedWith("Can only sell for asset");
       });
     });
-    describe("long by starting an EasyAuction", async () => {
+    describe("short by starting an EasyAuction", async () => {
       let auctionDeadline: number;
 
-      it("should be able to start an auction", async () => {
-        const auctionUSDCBalanceBefore = await usdc.balanceOf(auction.address);
-        const minBuy = 10 * 1e8;
-        const premium = 2000 * 1e6; // amount usdc paying
+      it("should be able to mint and start an auction phase", async () => {
+        const collateralAmount = utils.parseUnits("10");
+        const auctionOtokenBalanceBefore = await otoken1.balanceOf(auction.address);
+        const mintAmount = 10 * 1e8;
+        const sellAmount = 5 * 1e8;
+        const minPremium = utils.parseUnits("1"); // 1 eth min premium
 
         const blockNumber = await provider.getBlockNumber();
         const block = await provider.getBlock(blockNumber);
         const currentTimestamp = block.timestamp;
         auctionDeadline = currentTimestamp + 86400 * 1;
 
-        const minimalBidAmountPerOrder = 50 * 1e6; // min bid each order: 50 USDC
+        const minimalBidAmountPerOrder = 0.1 * 1e8; // min bid each order: 0.1 otoken
         const minFundingThreshold = 0;
 
-        await action.connect(owner).startAuction(
+        await action.connect(owner).mintAndStartAuction(
+          collateralAmount,
+          mintAmount,
+          sellAmount,
           auctionDeadline, // order cancel deadline
           auctionDeadline,
-          premium,
-          minBuy,
+          minPremium,
           minimalBidAmountPerOrder,
           minFundingThreshold,
           false
         );
-        const auctionUSDCBalanceAfter = await usdc.balanceOf(auction.address);
-        expect(auctionUSDCBalanceAfter.sub(auctionUSDCBalanceBefore).eq(premium)).to.be.true;
+        const auctionOtokenBalanceAfter = await otoken1.balanceOf(auction.address);
+        expect(auctionOtokenBalanceAfter.sub(auctionOtokenBalanceBefore).eq(sellAmount)).to.be.true;
       });
 
-      it('can long by participate in a "otoken selling auction"', async () => {
+      it("should start another auction with otoken left in the contract", async () => {
+        const collateralAmount = 0;
+        const auctionOtokenBalanceBefore = await otoken1.balanceOf(auction.address);
+        const mintAmount = 0;
+        const sellAmount = 5 * 1e8;
+        const minPremium = utils.parseUnits("1"); // 1 eth min premium
+
+        const blockNumber = await provider.getBlockNumber();
+        const block = await provider.getBlock(blockNumber);
+        const currentTimestamp = block.timestamp;
+        auctionDeadline = currentTimestamp + 86400 * 1;
+
+        const minimalBidAmountPerOrder = 0.1 * 1e8; // min bid each order: 0.1 otoken
+        const minFundingThreshold = 0;
+
+        await action
+          .connect(owner)
+          .mintAndStartAuction(
+            collateralAmount,
+            mintAmount,
+            sellAmount,
+            auctionDeadline,
+            auctionDeadline,
+            minPremium,
+            minimalBidAmountPerOrder,
+            minFundingThreshold,
+            false
+          );
+        const auctionOtokenBalanceAfter = await otoken1.balanceOf(auction.address);
+        expect(auctionOtokenBalanceAfter.sub(auctionOtokenBalanceBefore).eq(sellAmount)).to.be.true;
+      });
+
+      it('can short by participate in a "buy otoken auction"', async () => {
         const blockNumber = await provider.getBlockNumber();
         const block = await provider.getBlock(blockNumber);
         const currentTimestamp = block.timestamp;
         const auctionDeadline = currentTimestamp + 86400 * 1;
         // buyer create an auction to use 5 eth to buy 60 otokens
-        const sellAmount = 100 * 1e8;
-        const minPremium = 6000 * 1e6;
-        const seller = accounts[3];
-        const minBidPerOrder = 1 * 1e8;
-
-        await otoken1.mint(seller.address, sellAmount);
-        await otoken1.connect(seller).approve(auction.address, sellAmount);
-        await auction.connect(seller).initiateAuction(
+        const buyer = accounts[3];
+        const sellAmount = utils.parseUnits("5");
+        await token.mint(buyer.address, sellAmount);
+        await token.connect(buyer).approve(auction.address, sellAmount);
+        await auction.connect(buyer).initiateAuction(
+          token.address,
           otoken1.address,
-          usdc.address,
           auctionDeadline,
           auctionDeadline,
           sellAmount,
-          minPremium, // min premium amount
-          minBidPerOrder, // minimumBiddingAmountPerOrder
+          60 * 1e8, // min buy amount
+          1e6, // minimumBiddingAmountPerOrder
           0, // minFundingThreshold
           false, // isAtomicClosureAllowed
           ethers.constants.AddressZero, // accessManagerContract
@@ -338,21 +410,25 @@ describe("LongAction: Buying Puts", function () {
 
         const auctionIdToParticipate = await auction.auctionCounter();
 
-        const minBuyAmount = 10 * 1e8;
-        const premiumToPay = 700 * 1e6;
+        // the action participate in the action
+        const collateralAmount = utils.parseUnits("5");
+        const mintAmount = 5 * 1e8;
+        const minPremium = utils.parseUnits("0.2");
 
-        const auctionUSDCBalanceBefore = await usdc.balanceOf(auction.address);
+        const auctionOtokenBalanceBefore = await otoken1.balanceOf(auction.address);
         await action
           .connect(owner)
-          .bidInAuction(
+          .mintAndBidInAuction(
             auctionIdToParticipate,
-            [minBuyAmount],
-            [premiumToPay],
+            collateralAmount,
+            mintAmount,
+            [minPremium],
+            [mintAmount],
             ["0x0000000000000000000000000000000000000000000000000000000000000001"],
             "0x00"
           );
-        const auctionUSDCBalanceAfter = await usdc.balanceOf(auction.address);
-        expect(auctionUSDCBalanceAfter.sub(auctionUSDCBalanceBefore).eq(premiumToPay)).to.be.true;
+        const auctionOtokenBalanceAfter = await otoken1.balanceOf(auction.address);
+        expect(auctionOtokenBalanceAfter.sub(auctionOtokenBalanceBefore).eq(mintAmount)).to.be.true;
       });
     });
     it("should not be able to commit next token", async () => {
@@ -372,32 +448,32 @@ describe("LongAction: Buying Puts", function () {
       await expect(action.connect(owner).closePosition()).to.be.revertedWith("!VAULT");
     });
     it("should be able to close the position", async () => {
-      const actionBalanceBefore = await usdc.balanceOf(action.address);
-      const redeemPayout = 4000 * 1e6;
-
-      // mock payout asset
-      await usdc.mint(pool.address, redeemPayout);
-      await controller.setRedeemPayout(usdc.address, redeemPayout);
+      const actionBalanceBefore = await token.balanceOf(action.address);
+      const settlePayout = utils.parseUnits("9"); // assume we can get back 9 eth
+      await controller.setSettlePayout(settlePayout);
 
       await action.connect(vault).closePosition();
-      const actionBalanceAfter = await usdc.balanceOf(action.address);
-      expect(actionBalanceAfter.sub(actionBalanceBefore).eq(redeemPayout)).to.be.true;
+      const actionBalanceAfter = await token.balanceOf(action.address);
+      expect(actionBalanceAfter.sub(actionBalanceBefore).eq(settlePayout)).to.be.true;
       expect((await action.state()) === ActionState.Idle).to.be.true;
     });
     it("should revert if calling mint in idle phase", async () => {
-      const premium = 5000 * 1e6;
-      const buyAmount = 20 * 1e8;
+      const collateral = utils.parseUnits("10");
+      const amountOTokenToMint = 10 * 1e8;
+      const premium = utils.parseUnits("1");
       const order = await getOrder(
         action.address,
-        usdc.address,
-        premium,
-        counterpartyWallet.address,
         otoken1.address,
-        buyAmount,
+        amountOTokenToMint,
+        counterpartyWallet.address,
+        token.address,
+        premium.toString(),
         swap.address,
         counterpartyWallet.privateKey
       );
-      await expect(action.connect(owner).tradeAirswapOTC(order)).to.be.revertedWith("!Activated");
+      await expect(
+        action.connect(owner).mintAndTradeAirSwapOTC(collateral, amountOTokenToMint, order)
+      ).to.be.revertedWith("!Activated");
     });
   });
 });
